storage: Refactor code for the Jobs panel
Nobody is using the "jobs manager" anymore.

Closes #8062
