+++ pkg/storaged/block-details.jsx
@@ -49,6 +49,6 @@ export class BlockDetails extends React.Component {
 
         var content = <Content.Block client={this.props.client} block={block}/>;
 
-        return <StdDetailsLayout jobs={this.props.jobs} header={header} content={content}/>;
+        return <StdDetailsLayout client={this.props.client} header={header} content={content}/>;
     }
 }
