+++ pkg/storaged/details.jsx
@@ -25,7 +25,7 @@ import { BlockDetails } from "./block-details.jsx";
 import { DriveDetails } from "./drive-details.jsx";
 import { VGroupDetails } from "./vgroup-details.jsx";
 import { MDRaidDetails } from "./mdraid-details.jsx";
-import { Jobs } from "./job-views.jsx";
+import { JobsPanel } from "./jobs-panel.jsx";
 
 const _ = cockpit.gettext;
 
@@ -45,7 +45,7 @@ export class StdDetailsLayout extends React.Component {
                         <div id="detail-content">
                             { this.props.content }
                         </div>
-                        <Jobs jobs={this.props.jobs}/>
+                        <JobsPanel client={this.props.client}/>
                     </div>
                 </div>
             );
@@ -60,7 +60,7 @@ export class StdDetailsLayout extends React.Component {
                         <div id="detail-content">
                             { this.props.content }
                         </div>
-                        <Jobs jobs={this.props.jobs}/>
+                        <JobsPanel client={this.props.client}/>
                     </div>
                 </div>
             );
@@ -85,7 +85,6 @@ class Details extends React.Component {
 
     render() {
         var client = this.props.client;
-        var jobs = this.props.jobs;
 
         function go_up(event) {
             if (!event || event.button !== 0)
@@ -101,22 +100,22 @@ class Details extends React.Component {
 
             if (drive) {
                 name = utils.drive_name(drive);
-                body = <DriveDetails client={client} jobs={jobs} drive={drive}/>;
+                body = <DriveDetails client={client} drive={drive}/>;
             } else if (block) {
                 name = utils.block_name(block);
-                body = <BlockDetails client={client} jobs={jobs} block={block}/>;
+                body = <BlockDetails client={client} block={block}/>;
             }
         } else if (this.props.type == "vgroup") {
             var vgroup = client.vgnames_vgroup[this.props.name];
             if (vgroup) {
                 name = vgroup.Name;
-                body =  <VGroupDetails client={client} jobs={jobs} vgroup={vgroup}/>;
+                body =  <VGroupDetails client={client} vgroup={vgroup}/>;
             }
         } else if (this.props.type == "mdraid") {
             var mdraid = client.uuids_mdraid[this.props.name];
             if (mdraid) {
                 name = utils.mdraid_name(mdraid);
-                body =  <MDRaidDetails client={client} jobs={jobs} mdraid={mdraid}/>;
+                body =  <MDRaidDetails client={client} mdraid={mdraid}/>;
             }
         }
 
@@ -137,11 +136,11 @@ class Details extends React.Component {
     }
 }
 
-export function init(client, jobs) {
+export function init(client) {
     var page = document.getElementById("storage-detail");
 
     function show(type, name) {
-        React.render(<Details client={client} jobs={jobs} type={type} name={name}/>, page);
+        React.render(<Details client={client} type={type} name={name}/>, page);
         page.style.display = "block";
     }
