+++ pkg/storaged/mdraid-details.jsx
@@ -340,7 +340,7 @@ export class MDRaidDetails extends React.Component {
 
         var content = <Content.Block client={this.props.client} block={block}/>;
 
-        return <StdDetailsLayout jobs={this.props.jobs} alert={degraded_message}
+        return <StdDetailsLayout client={this.props.client} alert={degraded_message}
                                  header={ header }
                                  sidebar={ sidebar }
                                  content={ content }/>;
